backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
transformers==4.31.0
diffusers==0.21.4
accelerate==0.21.0
pillow==10.0.1
numpy==1.24.3
requests==2.31.0
safetensors==0.3.2
huggingface-hub==0.16.4
Pillow-stubs==0.2.0
